Fix redis 'disconnect' to not throw #24
I believe what's happening here is that sending the "QUIT" command
causes the redis server to close the connection, and then throws the
'DisconnectedException'. Throwing an exception in the withConnection
exception handler causes the original exception to not be rethrown.
Since this exception is expected, I've added code to instead ignore it.