runNMApp should finish cleanly when an empty string is received
@snoyberg This seems to resolve #37, where `DecodeFailure "demandInput:
not enough bytes"` is thrown.  Looks like what's happening here is that
the connection closes, and `appRead` yields `""`.  This is then passed
to the initial binary parser as a `Nothing`.  Since it isn't in the
middle of a parse, it throws an error since an empty string doesn't
store anything.

Am I correct in assuming `appRead` only returns empty string when the
connection is closed??