Bugfix: 2nd nmRead after closing didn't throw NMConnectionClosed
Instead, it blocked forever.  The fix is to check 'active' before
blocking on the channel.

There is more to this change than that, though.  This checking of active
means that we need to enforce the invariant that 'active' gets set to
'False' before (throwIO NMConnectionClosed) is enqueued.  Otherwise, it
would be possible (albeit highly unlikely) that nmRead is called twice
before that thread sees (active == False).  The second call would block
forever.