Ensure that every state is mentioned in input
Otherwise, we need some semantics for states that aren't mentioned in
the input. There seems to be an arbitrary choice here between keeping
the unmentioned states and dropping them. I'd lean towards dropping
them, in order to avoid memory leaks. To force the user to be explicit,
they need to specify every StateId in their input.