Temporary workaround: fix store to 7b56b3.
The lower bounds on dependencies are incompatible with packages in LTS-3
and LTS-5 that we need.  In order to be able to build this package
again, I've fixed the store version to the commit just before those
bounds were introduced.

In the long run, I think we should get rid of the superfluous lower
bounds in store again.