Changes to StaleKeySpec.
- Expectations are more precise
  length xs `shouldBe` 1
    -->
  xs `shouldBe` [x]

- Use race instead of race_, so no types are discarded.

- Get closer to the case that we actually want to model: a worker that
  - fails its heartbeat
  - then, takes on another job
  - then, dies for good.

I also changed checkStaleKeys again, to use threadDelay as a first
action, instead of after the check.  The reason is that otherwise, the
tests would occasionally fail:  To make sure that a job that was taken
by a worker after its heartbeat was failed was actually rescheduled, we
first check that the worker took the job.  But if the client starts
checking for stale keys right away, this check can fail (if the job was
already rescheduled).