Reduce code duplication in tests.
Introduces a function `raceAgainstVoids` that performs some action
concurrently to a list of non-terminating actions, killing the latter
when the former terminates.

This is a recurring pattern in the tests, where we often spawn some
workers and perform a test while they run.  This function allows us to
make sure every client is killed when the test finishes (which is
important, since lingering clients from former tests can lead to
test failures).