Don't force results and states in runSlave.
In order to make sure there were no Exceptions hidden in thunks that
could escape `tryAny`, the new states and output was fully evaluated.
This turned out to be very bad for performance.  In particular,
evaluating the `HashMap` of states was costly for performance.

This commit removes the `force`, and replaces the `tryAny` by a `catch`
encapsulating the whole function.