Assert `activeKey` has correct length.
We never want thelist stored under an `activeKey` to have a length
different from 0 or 1 (it models a `Maybe RequestId`, using a list in
order to have function like `rpoplpush`).

This commit adds a function that checks this invariant, and a call to
this function after every operation that adds an element to the
`activeKey` (currently only 1 invocation).

We don't need to check after removing an element, since the list
trivially has a minimal length of 0.

Closed #173.