Fix test: gets all slaves available (short).
The time the master waited for slaves was too short (for poor old
travis).  I've significantly increased the wait time.  In order to
finish the test in a timely manner regardless, I've also set an upper
bound on the number of slaves the master accepts, so that waiting stops
after all slaves have connected.  Since the bound is set to number of
available slaves, this does not decrease the power of the test.

Closed #166.