Reduce code duplication.
There were three parts in the code that re-enqueue the request of a
given worker: the worker itself, the heartbeat checker, and the stale
key checker.

This commit moves the re-enque logic into one function that is called by
the worker, hearbeat-, and stale key checker, increasing
maintainability (in particular, it will make it easier and safer to
include a "high-priority" request queue).