Fix another intermittent test failure.
In the test "Handles urgent jobs first", two jobs are submitted, one
urgent, and one normal.  The test waits, via Async.race, for the first
job to finish, which should be the urgent one.  However, both jobs
finish immediately, so the worker has a chance to finish both.  This
results in a race condition where completion of the non-urgent request
can be detected first (even though the urgent test completed earlier),
which makes the test fail.  The solution is to add a small delay to the
jobs.